AP-18416: Not use ThreadPool for Conda stdout/err listener
When using the KNIMEConstants.GLOBAL_THREAD_POOL we could end up in a
deadlock:
* The tread pool is full of threads waiting on Python installation tests
* The installation tests need to call Conda for finding the directory of
  the environment
* The Conda process starts to execute but the stdout and stderr
  listeners end up waiting forever because the thread pool is full

Before the extraction of the Conda plugin this was unlikely to happen
because the python2 plugin was activated (and the installation tests
were executed) when any workflow was loaded (because of the flow
variable). It was unlikely that the thread pool was full of threads
waiting for the Python installation tests.
Now, the python2 plugin is only activated when a node from the plugin is
loaded and therefore this is more likely to happen.

AP-18416 (Split repositories: knime-python2, knime-python3 and knime-conda)