AP-18784: Remove Micromamba PluginActivator to prevent broken installs
In case the micromamba execution check failed during feature installation
in KNIME, the plugin was installed but not the feature which prevented plugin
deinstallation but also did not update the plugin when installing the feature.

The Activator only contained a check whether we could run the micromamba executable,
but doing that during installation should fail the installation completely, so
we removed the check.

AP-18784 (Create environment during plugin installation)