AP-24511: Do not call pixi init but install from existing pixi.lock
Pixi init -i env.yaml does not run without an internet connection. The
lockfile is already part of the plugin but it does not have the correct
paths. Therefore, we load the lockfile, modify it, and finally install
the environment from the modified lockfile. This ensures that pixi does
not need to make any requests to online resources.

AP-24511 (Offline install with pixi does not work)

Co-authored-by: Benjamin Wilhelm <benjamin.wilhelm@knime.com>